dev/scripts/requirements.txt
# Requirements for font showcase generator
Pillow>=9.0.0
